feat: Update code to Subscription service
- Move code for the subscription service.
- Fixed issue with subscription order creation
