+++ processor/src/services/stripe-payment.service.ts
@@ -4,7 +4,6 @@ import {
   Cart,
   ErrorInvalidOperation,
   healthCheckCommercetoolsPermissions,
-  Money,
   statusHandler,
 } from '@commercetools/connect-payments-sdk';
 import {
@@ -40,7 +39,6 @@ import { stripeApi, wrapStripeError } from '../clients/stripe.client';
 import { log } from '../libs/logger';
 import { StripeEventConverter } from './converters/stripeEventConverter';
 import { convertPaymentResultCode } from '../utils';
-import { SubscriptionEventConverter } from './converters/subscriptionEventConverter';
 import { CtPaymentCreationService } from './ct-payment-creation.service';
 import { stripeCustomerIdFieldName } from '../custom-types/custom-types';
 import { StripeCustomerService } from './stripe-customer.service';
@@ -52,14 +50,12 @@ import { CartUpdateAction } from '@commercetools/platform-sdk';
 
 export class StripePaymentService extends AbstractPaymentService {
   private stripeEventConverter: StripeEventConverter;
-  private subscriptionEventConverter: SubscriptionEventConverter;
   private customerService: StripeCustomerService;
   private paymentCreationService: CtPaymentCreationService;
 
   constructor(opts: StripePaymentServiceOptions) {
     super(opts.ctCartService, opts.ctPaymentService, opts.ctOrderService);
     this.stripeEventConverter = new StripeEventConverter();
-    this.subscriptionEventConverter = new SubscriptionEventConverter();
     this.customerService = new StripeCustomerService(opts.ctCartService);
     this.paymentCreationService = new CtPaymentCreationService({
       ctCartService: opts.ctCartService,
@@ -565,96 +561,10 @@ export class StripePaymentService extends AbstractPaymentService {
 
       if (event.type === StripeEvent.PAYMENT_INTENT__SUCCEEDED) {
         const ctCart = await this.ctCartService.getCartByPaymentId({ paymentId: updateData.id });
-        const updatedCart = await this.updateCartAddress(event, ctCart);
+        const { latest_charge } = event.data.object as Stripe.PaymentIntent;
+        const charge = await stripeApi().charges.retrieve(latest_charge as string);
+        const updatedCart = await this.updateCartAddress(charge, ctCart);
         await this.createOrder({ cart: updatedCart, paymentIntentId: updateData.pspReference });
-      }
-    } catch (e) {
-      log.error('Error processing notification', { error: e });
-      return;
-    }
-  }
-
-  /**
-   * Retrieves modified payment data based on the given Stripe event for subscriptions.
-   *
-   * @param {Stripe.Event} event - The Stripe event object to extract data from.
-   * @return {ModifyPayment} - An object containing modified payment data.
-   */
-  public async processSubscriptionEvent(event: Stripe.Event): Promise<void> {
-    log.info('Processing subscription notification', { event: JSON.stringify(event.id) });
-    try {
-      const dataInvoice = event.data.object as Stripe.Invoice;
-      const invoiceExpanded = await this.paymentCreationService.getStripeInvoiceExpanded(dataInvoice.id);
-      const subscription = invoiceExpanded.subscription as Stripe.Subscription;
-      const invoicePaymentIntent = invoiceExpanded.payment_intent as Stripe.PaymentIntent;
-
-      let payment = await this.ctPaymentService.getPayment({
-        id: subscription.metadata?.ct_payment_id ?? '',
-      });
-      if (!payment) {
-        log.error(`Cannot process invoice with ID: ${dataInvoice.id}. Missing Payment can be trial days.`);
-        return;
-      }
-
-      const failedPaymentIntent = await this.ctPaymentService.findPaymentsByInterfaceId({
-        interfaceId: invoicePaymentIntent.id,
-      });
-      const isPaymentFailed = failedPaymentIntent.length > 0;
-      if (failedPaymentIntent.length > 0) {
-        //Update a failed payment if it has a Failed transaction
-        payment = failedPaymentIntent[0];
-      }
-      const isPaymentChargePending = this.ctPaymentService.hasTransactionInState({
-        payment,
-        transactionType: PaymentTransactions.CHARGE,
-        states: [PaymentStatus.PENDING],
-      });
-
-      const updateData = this.subscriptionEventConverter.convert(
-        event,
-        invoiceExpanded,
-        isPaymentChargePending,
-        payment,
-      );
-
-      if (!isPaymentChargePending && !isPaymentFailed) {
-        log.info(`Subscription Payment ${payment} do not have Transaction in pending state`);
-        const eventCartId = dataInvoice.subscription_details?.metadata?.cart_id;
-        if (!eventCartId) {
-          log.error(`Cannot process invoice with ID: ${dataInvoice.id}. Missing cart.`);
-          return;
-        }
-
-        const cart = await this.ctCartService.getCart({ id: eventCartId });
-        //If it is invoice.payment_failed the amount is the amount_due
-        const isInvoicePaid = event.type.startsWith('invoice.paid');
-        const amountPlanned: Money = {
-          currencyCode: dataInvoice.currency.toUpperCase(),
-          centAmount: isInvoicePaid ? dataInvoice.amount_paid : dataInvoice.amount_due,
-        };
-        const createdPayment = await this.paymentCreationService.handleCtPaymentSubscription({
-          cart,
-          amountPlanned,
-          interactionId: updateData.pspReference,
-        });
-
-        await this.addPaymentToOrder(payment.id, createdPayment);
-        updateData.id = createdPayment;
-      }
-
-      for (const tx of updateData.transactions) {
-        const updatedPayment = await this.ctPaymentService.updatePayment({
-          ...updateData,
-          transaction: tx,
-        });
-
-        log.info('Subscription payment updated after processing the notification', {
-          paymentId: updatedPayment.id,
-          version: updatedPayment.version,
-          pspReference: updateData.pspReference,
-          paymentMethod: updateData.paymentMethod,
-          transaction: JSON.stringify(tx),
-        });
       }
     } catch (e) {
       log.error('Error processing notification', { error: e });
@@ -696,9 +606,7 @@ export class StripePaymentService extends AbstractPaymentService {
     }
   }
 
-  public async updateCartAddress(event: Stripe.Event, ctCart: Cart): Promise<Cart> {
-    const { latest_charge } = event.data.object as Stripe.PaymentIntent;
-    const charge = await stripeApi().charges.retrieve(latest_charge as string);
+  public async updateCartAddress(charge: Stripe.Charge, ctCart: Cart): Promise<Cart> {
     const { billing_details, shipping } = charge;
     let billingAlias: Stripe.Charge.BillingDetails | Stripe.Charge.Shipping;
     if (!shipping) {
